refactor(category): CategoryTemplateApplicationServiceTest 삭제
